[room] calendarDate 값 초기화 기능 추가
