+++ src/app/(service)/(home)/home/page.tsx
@@ -1,6 +1,6 @@
 'use client'
-import { useState } from 'react'
-import { useRecoilValue } from 'recoil'
+import { useEffect, useState } from 'react'
+import { useRecoilValue, useResetRecoilState, useSetRecoilState } from 'recoil'
 
 import DaySchedule from '@/components/main/DaySchedule'
 import { calendarDateState } from '@/lib/state/calendarDateState'
@@ -9,7 +9,19 @@ import Calendar from '@/components/common/calendar/Calendar'
 
 export default function MainPageDay() {
   const { width, height } = useWindowSize()
+  const currentDate = new Date()
+  const setCalendarDate = useSetRecoilState(calendarDateState)
   const calendarDate = useRecoilValue(calendarDateState)
+
+  useEffect(() => {
+    return () => {
+      setCalendarDate({
+        year: currentDate.getFullYear(),
+        month: currentDate.getMonth(),
+        date: currentDate.getDate()
+      })
+    }
+  }, [])
 
   return (
     <>
